services/ai-service/requirements.txt
# AI Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
kafka-python==2.0.2
openai==1.3.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.35.0
torch==2.1.0
pgvector==0.2.4
python-dotenv==1.0.0
httpx==0.25.2
langchain==0.0.350
langchain-community==0.0.10
chromadb==0.4.18
faiss-cpu==1.7.4